backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "novex-backend"
version = "0.1.0"
description = "Novex backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "pydantic>=2.8,<3.0",
    "pydantic-settings>=2.4,<3.0",
    "sqlalchemy>=2.0,<3.0",
    "psycopg2-binary>=2.9,<3.0",
    "alembic>=1.13,<2.0",
    "redis>=5.0,<6.0",
    "httpx>=0.27,<1.0",
    "python-dotenv>=1.0,<2.0",
    "PyJWT>=2.8,<3.0",
    "email-validator>=2.0,<3.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "ruff>=0.6,<1.0",
    "black>=24.0,<25.0",
    "mypy>=1.10,<2.0"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "workers*", "integrations*"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"